Serve a single page application
In the default mode, only files and folders are served. However, to serve a SPA, all paths that do not match a file have to be forwarded to index.html. Your SPA will then render the appropriate view depending on the path.

This commit turns on the SPA mode by default.